cincinnati: Expose dead-end release reason via MOTD
After checking whether the current Cincinnati node is a dead-end or
not, write out a new MOTD file at `/run/zincati/public/motd.d/deadend.motd`
reflecting the dead-end status. If the current release is a deadend,
a message is printed in the MOTD containing the reason for the
release being a deadend. If not, a blank MOTD is written.

To make the final write atomic, a tempfile is written within the
`/run/zincati/public/motd.d/` directory, and renamed to the location
of the complete MOTD to display. By writing the tempfile in the same
directory as the final MOTD first, the tempfile inherits the same
attributes, such as SELinux labels, as the final location should
have. This is particularly important as sshd requires a specific
label for any MOTD files it reads.

Entries to tmpfiles.d are added to 1) create the Zincati-owned motd.d
directory and 2) create a symlink out of `/run/motd.d/` to allow
`pam_motd` to find the final MOTD file to display upon SSH login.

The reason for using a symlink from /run/motd.d/ rather than writing
to a file in /run/motd.d/ directly is that Zincati runs as the
`zincati` sysuser, and does not have permissions to modify the
root-owned `/run/motd.d/`, which would be required for the atomic
write operations (e.g. renaming a file to under `/run/motd.d/`).

Resolves: #90
Co-authored by: Sohan Kunkerkar <sohank2602@gmail.com>